update the package version for the task managers
