+++ src/subscriber/Services/Queues/Azure/AzureServiceBusClient.cs
@@ -1,13 +1,7 @@
 using subscriber.Services.Queues.Exceptions;
 using Azure.Messaging.ServiceBus;
 using Azure.Messaging.ServiceBus.Administration;
-using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace subscriber.Services.Queues.Azure;
